Added functions to calculate compression defined as the sum of total encoding weights, decoding weights and transform memory resource counts over total synapses.
The file _spaun/spaun_main_wm_only.py is called only by build_spaun_wm_only.py and is used to construct only the WorkingMemory module.

The file build_spaun_all_modules.py is used to build a SPAUN model and run the preprocessing function on it.

The file build_spaun_wm_only.py is used to build a SPAUN model with only the WorkingMemmory module and run preprocessing on it.

The file calc_synapses_and_fanout.py will build and run preprocessing on either the entire SPAUN or SPAUN with only the WorkingMemory module
and then compute the compression ratio and associated data.

Finally, compute_stats.py has functionality to compute the compression ratio and associated data.

This branch does not need to go to mainline.